src/pyproject.toml
[project]
name = "professional-me-ocr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
'easyocr',
'pytesseract',
'tesseract',
'transformers',
'torch',
'Pillow',
'numpy',
'Levenshtein',
'loguru'
]
